add npi , first modifier and total charge to cms

+++ src/main/java/com/cob/billing/model/clinical/patient/CPTCode.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 public class CPTCode {
     private String serviceCode;
     private String modifier;
-    private Float unit;
-    private Integer charge;
+    private Integer unit;
+    private Float charge;
 }
